same for firestore functions
